@@ -1,8 +1,7 @@
-﻿namespace XTI_GitHub
+﻿namespace XTI_GitHub;
+
+public sealed class FilterIssueRequest
 {
-    public sealed class FilterIssueRequest
-    {
-        public bool IncludeOpenOnly { get; set; }
-        public int? Milestone { get; set; }
-    }
+    public bool IncludeOpenOnly { get; set; }
+    public int? Milestone { get; set; }
 }
