@@ -1,9 +1,8 @@
 ﻿using System;
 
-namespace XTI_GitHub
+namespace XTI_GitHub;
+
+public sealed record GitHubMilestone(int Number, string Title, string State = "Open")
 {
-    public sealed record GitHubMilestone(int Number, string Title, string State = "Open")
-    {
-        public bool IsOpen() => State.Equals("Open", StringComparison.OrdinalIgnoreCase);
-    }
+    public bool IsOpen() => State.Equals("Open", StringComparison.OrdinalIgnoreCase);
 }
