@@ -1,4 +1,3 @@
-﻿namespace XTI_GitHub
-{
-    public sealed record GitHubRelease(int ID, string TagName, GitHubReleaseAsset[] Assets);
-}
+﻿namespace XTI_GitHub;
+
+public sealed record GitHubRelease(int ID, string TagName, GitHubReleaseAsset[] Assets);
