@@ -1,4 +1,3 @@
-﻿namespace XTI_GitHub
-{
-    public sealed record GitHubReleaseAsset(int ID, string Name, string ContentType, string Url);
-}
+﻿namespace XTI_GitHub;
+
+public sealed record GitHubReleaseAsset(int ID, string Name, string ContentType, string Url);
