@@ -6,22 +6,21 @@ using XTI_GitHub;
 using XTI_GitHub.Web;
 using XTI_Secrets.Extensions;
 
-namespace XTI_Git.IntegrationTests
+namespace XTI_Git.IntegrationTests;
+
+internal static class TestExtensions
 {
-    public static class TestExtensions
+    public static void AddTestServices(this IServiceCollection services, IHostEnvironment hostEnv, string repoOwner, string repoName, string gitRepoPath)
     {
-        public static void AddTestServices(this IServiceCollection services, IHostEnvironment hostEnv, string repoOwner, string repoName, string gitRepoPath)
+        services.AddFileSecretCredentials(hostEnv);
+        services.AddSingleton<XtiFolder>();
+        services.AddScoped<XtiGitHubRepository>(sp =>
         {
-            services.AddFileSecretCredentials(hostEnv);
-            services.AddSingleton<XtiFolder>();
-            services.AddScoped<XtiGitHubRepository>(sp =>
-            {
-                return new WebXtiGitHubRepository(repoOwner, repoName);
-            });
-            services.AddScoped<XtiGitRepository>(sp =>
-            {
-                return new GitLibXtiGitRepository(gitRepoPath);
-            });
-        }
+            return new WebXtiGitHubRepository(repoOwner, repoName);
+        });
+        services.AddScoped<IXtiGitRepository>(sp =>
+        {
+            return new GitLibXtiGitRepository(gitRepoPath);
+        });
     }
 }
